pyproject.toml
[project]
name = "nomad"
version = "0.1.0"
description = "Multi-agent system for trip planning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.8.6",
    "fast-flights>=2.2",
    "flights>=0.8.4",
    "httpx>=0.27",
    "langchain>=1.2.15",
    "langchain-ollama>=0.2.0",
    "langgraph>=1.1.9",
    "langgraph-cli>=0.4.24",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
]

[project.scripts]
nomad = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
